feature(Tests): expose `tuneability_test` to other dependent libraries
